requirements.txt
openai
python-dotenv
requests
beautifulsoup4
PyMuPDF
pandas
selenium